Split DaemonClient.swift along its six concerns
Step #8 of the architectural roadmap. `DaemonClient.swift` was 501
lines covering six grown-organically concerns: auto-spawn, version
handshake, restart-lock acquisition, stderr log forwarding, StallTimer
wiring, and socket-readiness polling. The seams already existed in
section comments — this PR moves them to discrete files via
`extension DaemonClient`.

After:

  • `DaemonClient.swift` (328 lines) — public surface (`connect`,
    `withDaemonClient`), `openClient` initialize handshake, daemon
    spawn + socket-readiness poll, `DaemonClientError` enum.
  • `DaemonRestart.swift` (179 lines) — version-mismatch detection,
    restart-lock acquisition, kill-and-wait, `versionsMatch` /
    `baseVersion` / `gitDescribeRange` git-describe parsing.
  • `DaemonClientChannel.swift` (55 lines) — notification-handler
    wiring (stderr log forwarder + stall-timer bumpers).

The header docstring on `DaemonClient` describes the file layout so
future maintainers don't have to grep for which file owns what.

`private static func` becomes `static func` (default internal) at the
extension boundary. `internal` is module-scoped, but PreviewsCLI is
the executable target — internal IS effectively private from any
external observer.

No behavior change. AGENTS.md:240-241 documents subtleties (timer seed,
debug logging level) that all live in `withDaemonClient` — preserved
verbatim.

Acceptance per the plan:
  • `swift build` clean.
  • `VersionHandshakeTests` passes — the version-mismatch path
    (`_PREVIEWSMCP_TEST_DAEMON_VERSION` env override + restart) still
    triggers a clean restart through `restartDaemonAndReconnect`.
  • `MacOSMCPTests` (7 tests, including hot-reload, heartbeat, and
    `withTimeout` pthread timer) all pass.
